fix: blog category metadata error
